Opravy

+++ Source/Orts.Simulation/Simulation/RollingStocks/MSTSLocomotive.cs
@@ -1494,7 +1494,7 @@ namespace Orts.Simulation.RollingStocks
             Heating_OffOn = inf.ReadBoolean();
             SwitchingVoltageMode_OffAC = inf.ReadBoolean();
             SwitchingVoltageMode_OffDC = inf.ReadBoolean();
-            SwitchingVoltageMode = inf.ReadInt16();
+            SwitchingVoltageMode = inf.ReadInt32();
             THeating = inf.ReadSingle();
 
             base.Restore(inf);
